requirements.txt
ultralytics>=8.0.0
openai-whisper>=20231117
pandas>=2.0.0
opencv-python-headless>=4.8.0
ffmpeg-python>=0.2.0
torch>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

